requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
